break out github api stuff to separate module
(SOFTWARE-3564)
