+++ src/webhook_app.py
@@ -1,7 +1,6 @@
 """
 Application File
 """
-import base64
 import flask
 import flask.logging
 from   flask import Flask, Response, request
@@ -13,9 +12,9 @@ import re
 import subprocess
 from   subprocess import PIPE
 import sys
-import urllib.request
 
 from webapp import default_config
+from webapp import github
 from webapp.common import run_git_cmd
 from webapp.models import GlobalData
 
@@ -72,55 +71,13 @@ if not webhook_secret:
     app.logger.warning("Note, no WEBHOOK_SECRET_KEY configured; "
                        "GitHub payloads will not be validated.")
 
-def mk_github_authstr(user, passwd):
-    return base64.encodestring('%s:%s' % (user,passwd)).replace('\n', '')
-
 gh_api_user = global_data.webhook_gh_api_user
 gh_api_token = _readfile(global_data.webhook_gh_api_token)
-if gh_api_token:
-    gh_api_authstr = mk_github_authstr(gh_api_user, gh_api_token)
+if gh_api_user and gh_api_token:
+    github.api_setup(gh_api_user, gh_api_token)
 else:
-    gh_api_authstr = None
     app.logger.warning("Note, no WEBHOOK_GH_API_TOKEN configured; "
                        "GitHub comments will not be published.")
-
-gh_api_baseurl = "https://api.github.com"
-
-def github_api_call(method, url, data):
-    req = urllib.request.Request(url, data)
-    add_auth_header(req)
-    add_gh_preview_header(req)
-    req.get_method = lambda : method
-    resp = urllib.request.urlopen(req)
-    # resp headers in: resp.headers
-    # resp body in resp.read()
-    # for extended responses, follow resp.headers.getheader('link') -> next
-    return resp
-
-def add_auth_header(req):
-    if gh_api_authstr:
-        req.add_header("Authorization", "Basic %s" % gh_api_authstr)
-
-def github_api_path2url(api_path, **kw):
-    fmtstr = re.sub(r':([a-z]+)\b', r'{\1}', api_path)
-    path = fmtstr.format(**kw)
-    url = "%s/%s" % (gh_api_baseurl, path)
-    return url
-
-def publish_issue_comment(owner, repo, num, body):
-    api_path = "/repos/:owner/:repo/issues/:number/comments"
-    url = github_api_path2url(api_path, owner=owner, repo=repo, number=num)
-    data = {'body': body}
-    resp = github_api_call('POST', url, data)
-    return resp
-
-def publish_pr_review(owner, repo, num, body, action, sha):
-    # action: APPROVE, REQUEST_CHANGES, or COMMENT
-    api_path = "/repos/:owner/:repo/pulls/:number/reviews"
-    url = github_api_path2url(api_path, owner=owner, repo=repo, number=num)
-    data = {'body': body, 'event': action, commit_id: sha}
-    resp = github_api_call('POST', url, data)
-    return resp
 
 def validate_webhook_signature(data, x_hub_signature):
     if webhook_secret:
